pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["mochuelo-kun <mochuelokun@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# langchain = "^0.1.1"
# pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
interlab = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "src.run:start"
# context_query = "notebooks.context_query_poc:start"
